seperate keymap models from class
